Secureproject

+++ server/app.py
@@ -20,7 +20,7 @@ class Signup(Resource):
         db.session.commit()
         session["user_id"] = user.id
 
-        return user.to_dict(rules=('-stores', '-purchases')), 201
+        return user.to_dict(), 201
     
     except IntegrityError:
        return {"error": "Username already exist"}, 422
@@ -32,9 +32,9 @@ class CheckSession(Resource):
    def get(self):
       user = User.query.filter(User.id == session.get('user_id')).first()
       if user:
-         return user.to_dict(rules=('-stores', '-purchases')), 200
+         return user.to_dict(), 200
       else:
-         return {"message" : "Not Authorized"}, 401
+         return {"error" : "Not Authorized"}, 401
 
 class Login(Resource):
    def post(self):
@@ -44,77 +44,88 @@ class Login(Resource):
       user = User.query.filter(User.username == username).first()
       if user and user.authenticate(password):
          session["user_id"] = user.id
-         return user.to_dict(rules=('-stores', '-purchases')), 200
+         return user.to_dict(), 200
       else:
-         return {"errors": "Incorrect username or password"}, 422
+         return {"error": "Incorrect username or password"}, 422
       
 class Logout(Resource):
    def delete(self):
       session["user_id"] = None
-      return {"message": "Log out"}, 204
+      return {"error": "Log out"}, 204
    
-class StoreList(Resource):
+class Stores(Resource):
    def get(self):
       user = User.query.filter(User.id == session.get('user_id')).first()
       if user:
          stores = Store.query.all()
-         stores_dict = [store.to_dict(rules=('-owner', '-items')) for store in stores]
+         stores_dict = [store.to_dict(rules=('-owner',)) for store in stores]
          return stores_dict, 200
       else:
-         return {"message" : "Not Authorized"}, 401
+         return {"error" : "Not Authorized"}, 401
       
    def post(self):
-      try:
-         data = request.get_json()
-         store_name = data.get("store_name")
-         description = data.get("description")
-         img_url = data.get("img_url")
-         user_id = session.get('user_id')
+      user = User.query.filter(User.id == session.get('user_id')).first()
+      if user.isAdmin:
+         try:
+            data = request.get_json()
+            store_name = data.get("store_name")
+            description = data.get("description")
+            img_url = data.get("img_url")
+            user_id = session.get('user_id')
 
-         new_store = Store(store_name=store_name, description=description, 
-                           img_url=img_url, user_id=user_id)
+            new_store = Store(store_name=store_name, description=description, 
+                              img_url=img_url, user_id=user_id)
+            
+            db.session.add(new_store)
+            db.session.commit()
+
+            return new_store.to_dict(rules=('-owner', )), 201
          
-         db.session.add(new_store)
-         db.session.commit()
-
-         return new_store.to_dict(rules=('-owner', '-items')), 201
+         except IntegrityError:
+            return {"error": "Store name already exist"}, 422
       
-      except IntegrityError:
-         return {"error": "Store name already exist"}, 422
+      else:
+         return {"error" : "Not Authorized"}, 401
       
 class StoreById(Resource):
    def get(self, storeid):
-      store = Store.query.filter_by(id=storeid).first()
-      if store:
-         return store.to_dict(rules=('-owner', )), 200
+      user = User.query.filter(User.id == session.get('user_id')).first()
+      if user:
+         store = Store.query.filter_by(id=storeid).first()
+         if store:
+            return store.to_dict(rules=('-owner', )), 200
+         else:
+            return {}, 404
       else:
-         return {}, 404
+         return {"error" : "Not Authorized"}, 401
    
    def delete(self, storeid):
-      store = Store.query.filter_by(id=storeid).first()
-      db.session.delete(store)
-      db.session.commit()
-      return {}, 204
+      user = User.query.filter(User.id == session.get('user_id')).first()
+      if user.isAdmin:
+         store = Store.query.filter_by(id=storeid).first()
+         db.session.delete(store)
+         db.session.commit()
+         return {}, 204
+      else:
+         return {"error" : "Not Authorized"}, 401
    
    def patch(self, storeid):
-      store = Store.query.filter_by(id=storeid).first()
-      try:
-         data = request.get_json()
-         for attr in data:
-            setattr(store, attr, data.get(attr))
+      user = User.query.filter(User.id == session.get('user_id')).first()
+      if user.isAdmin:
+         store = Store.query.filter_by(id=storeid).first()
+         try:
+            data = request.get_json()
+            for attr in data:
+               setattr(store, attr, data.get(attr))
 
-         db.session.add(store)
-         db.session.commit()
-         return store.to_dict(rules=('-owner', )), 200
-      
-      except IntegrityError:
-         return {"error": "Store name already exist"}, 422
-      
-class StoresByOwner(Resource):
-   def get(self, userid):
-      stores = Store.query.filter_by(user_id = userid).all()
-      stores_dict = [store.to_dict() for store in stores]
-      return stores_dict, 200
+            db.session.add(store)
+            db.session.commit()
+            return store.to_dict(), 200
+         
+         except IntegrityError:
+            return {"error": "Store name already exist"}, 422
+      else:
+         return {"error" : "Not Authorized"}, 401
 
    
 class Items(Resource):
@@ -122,51 +133,70 @@ class Items(Resource):
       user = User.query.filter(User.id == session.get('user_id')).first()
       if user:
          items = Item.query.all()
-         items_dict = [item.to_dict(rules =('-store', )) for item in items]
+         items_dict = [item.to_dict() for item in items]
          return items_dict, 200
       else: 
-         return {"message" : "Not Authorized"}, 401
+         return {"error" : "Not Authorized"}, 401
       
    def post(self):
-      data = request.get_json()
-      name = data.get("name")
-      description = data.get("description")
-      img_url = data.get("img_url")
-      price = data.get("price")
-      quantity = data.get("quantity")
-      store_id = data.get("store_id")
+      user = User.query.filter(User.id == session.get('user_id')).first()
+      if user.isAdmin:
+         data = request.get_json()
+         name = data.get("name")
+         description = data.get("description")
+         img_url = data.get("img_url")
+         price = data.get("price")
+         quantity = data.get("quantity")
+         store_id = data.get("store_id")
 
-      newItem = Item(name=name, description=description,
-                     img_url=img_url, price=price,
-                     quantity=quantity, store_id=store_id)
-      
-      db.session.add(newItem)
-      db.session.commit()
-      return newItem.to_dict(rules =('-store', )), 201
+         newItem = Item(name=name, description=description,
+                        img_url=img_url, price=price,
+                        quantity=quantity, store_id=store_id)
+         
+         db.session.add(newItem)
+         db.session.commit()
+         return newItem.to_dict(), 201
+      else: 
+         return {"error" : "Not Authorized"}, 401
 
       
 class ItembyId(Resource):
    def get(self, itemid):
       item = Item.query.filter_by(id=itemid).first()
       if item:
-         return item.to_dict(rules =('-store', )), 200
+         return item.to_dict(), 200
       else:
          return {}, 404
       
    def delete(self, itemid):
+      user = User.query.filter(User.id == session.get('user_id')).first()
       item = Item.query.filter_by(id=itemid).first()
-      db.session.delete(item)
-      db.session.commit()
-      return {}, 204
+      if user.isAdmin and item.store in user.stores:
+         db.session.delete(item)
+         db.session.commit()
+         return {}, 204
+      else: 
+         return {"error" : "Not Authorized"}, 401
    
    def patch(self, itemid):
+      user = User.query.filter(User.id == session.get('user_id')).first()
       item = Item.query.filter_by(id=itemid).first()
       data = request.get_json()
-      for attr in data:
-            setattr(item, attr, data.get(attr))
-      db.session.add(item)
-      db.session.commit()
-      return item.to_dict(rules =('-store', )), 200
+      if user.isAdmin and item.store in user.stores:
+         for attr in data:
+               setattr(item, attr, data.get(attr))
+         db.session.add(item)
+         db.session.commit()
+         return item.to_dict(), 200
+      
+      elif user and len(data) == 1 and "quantity" in data:
+         setattr(item, "quantity", data.get("quantity") )
+         db.session.add(item)
+         db.session.commit()
+         return item.to_dict(), 200
+      
+      else: 
+         return {"error" : "Not Authorized"}, 401
    
 class Purchases(Resource):
    def get(self):
@@ -175,48 +205,34 @@ class Purchases(Resource):
       return purchases_dict, 200
    
    def post(self):
-      data = request.get_json()
-      user_id = session.get('user_id')
-      item_id = data.get("item_id")
-      quantity = data.get("quantity")
-      amt_spent = data.get("amt_spent")
+      user = User.query.filter(User.id == session.get('user_id')).first()
+      if user:
+         data = request.get_json()
+         user_id = session.get('user_id')
+         item_id = data.get("item_id")
+         quantity = data.get("quantity")
+         amt_spent = data.get("amt_spent")
 
-      newPurchase = Purchase(user_id=user_id, item_id=item_id, quantity=quantity, amt_spent=amt_spent)
-      db.session.add(newPurchase)
-      db.session.commit()
+         newPurchase = Purchase(user_id=user_id, item_id=item_id, quantity=quantity, amt_spent=amt_spent)
+         db.session.add(newPurchase)
+         db.session.commit()
+         
+         return newPurchase.to_dict(), 201
       
-      return newPurchase.to_dict(), 201
+      else: 
+         return {"error" : "Not Authorized"}, 401
 
 
-   
-class PurchasesByUser(Resource):
-   def get(self, userid):
-      purchases = Purchase.query.filter_by(user_id = userid).all()
-      purchases_dict = [purchase.to_dict(rules=('-user',)) for purchase in purchases]
-      return purchases_dict, 200
-   
-class PurchasesByStore(Resource):
-   def get(self, storeid):
-      store = Store.query.filter_by(id = storeid).first()
-      items = store.items
-      purchases = []
-      for item in items:
-         purchases.extend(Purchase.query.filter_by(item_id = item.id).all())
-      purchases_dict = [purchase.to_dict() for purchase in purchases]
-      return purchases_dict, 200
-   
+
 api.add_resource(Signup, '/api/signup')
 api.add_resource(CheckSession, '/api/check_session')
 api.add_resource(Login, '/api/login')
 api.add_resource(Logout, '/api/logout')
-api.add_resource(StoreList, '/api/stores')
+api.add_resource(Stores, '/api/stores')
 api.add_resource(StoreById, '/api/stores/<int:storeid>')
-api.add_resource(StoresByOwner, '/api/stores/owner/<int:userid>')
 api.add_resource(Items, '/api/items')
 api.add_resource(ItembyId, '/api/items/<int:itemid>')
-api.add_resource(PurchasesByStore, '/api/purchases/stores/<int:storeid>')
 api.add_resource(Purchases, '/api/purchases')
-api.add_resource(PurchasesByUser, '/api/purchases/<int:userid>')
 
 
 @app.route('/')
